// src/dtos/post/likeOrDislike.dto.ts
import z from "zod";

export interface LikeOrDislikePostInputDTO {
  postId: string;
  token: string;
  like: boolean;
}

export type LikeOrDislikePostOutputDTO = { message: "Reação enviada" };

export const LikeOrDislikePostSchema = z
  .object({
    postId: z.string().min(1),
    token: z.string().min(1),
    like: z.boolean(),
  })
  .transform((data) => data as LikeOrDislikePostInputDTO);